+++ BinksSwitch/Network/Entities/FirewallRule.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Net;
 using System.Net.NetworkInformation;
 using BinksSwitch.Annotations;
@@ -14,104 +15,145 @@ namespace BinksSwitch.Network.Entities
             Permit
         }
 
-        private const ushort AnyPort = 0;
-
         public Operation RuleOperation { get; set; }
-        public Direction? RuleDirection { get; set; }
+        [CanBeNull] public Direction? RuleDirection { get; set; }
         [CanBeNull] public PhysicalAddress SourceMac { get; set; }
         [CanBeNull] public IPAddress SourceIp { get; set; }
         [CanBeNull] public PhysicalAddress DestinationMac { get; set; }
         [CanBeNull] public IPAddress  DestinationIp{ get; set; }
-        public Protocol? RuleProtocol { get; set; }
-        public ushort SourcePort { get; set; }
-        public ushort DestinationPort { get; set; }
+        [CanBeNull] public Protocol? RuleProtocol { get; set; }
+        [CanBeNull] public ushort? SourcePort { get; set; }
+        [CanBeNull] public ushort? DestinationPort { get; set; }
+
+        private bool resolveValidationStack(Stack<Func<bool>> stack)
+        {
+            foreach (Func<bool> rule in stack)
+            {
+                if (!rule())
+                {
+                    return false;
+                }
+            }
+            return true;
+        }
 
         public bool IsMatch(EthernetPacket packet)
         {
-            if (SourceMac != null && !SourceMac.Equals(packet.SourceHardwareAddress))
+            var validationStack = new Stack<Func<bool>>();
+
+            if (SourceMac != null)
             {
-                return false;
+                validationStack.Push(() => SourceMac.Equals(packet.SourceHardwareAddress));
             }
 
-            if (DestinationMac != null && DestinationMac.Equals(packet.DestinationHardwareAddress))
+            if (DestinationMac != null)
             {
-                return false;
+                validationStack.Push(() => DestinationMac.Equals(packet.DestinationHardwareAddress));
             }
 
             var arp = packet.Extract<ArpPacket>();
             if (arp != null && RuleProtocol.Equals(Protocol.Arp))
             {
-                return true;
+                return resolveValidationStack(validationStack);
             }
 
             var lldp = packet.Extract<LldpPacket>();
             if (lldp != null && RuleProtocol.Equals(Protocol.Lldp))
             {
-                return true;
+                return resolveValidationStack(validationStack);
             }
+
 
             var ipv4 = packet.Extract<IPv4Packet>();
             if (ipv4 != null)
             {
-                if (SourceIp != null && !SourceIp.Equals(ipv4.SourceAddress))
+                if (SourceIp != null)
                 {
-                    return false;
+                    validationStack.Push(() => SourceIp.Equals(ipv4.SourceAddress));
                 }
 
-                if (DestinationIp != null && DestinationIp.Equals(ipv4.DestinationAddress))
+                if (DestinationIp != null)
                 {
-                    return false;
+                    validationStack.Push(() => DestinationIp.Equals(ipv4.DestinationAddress));
                 }
 
                 if (RuleProtocol.Equals(Protocol.Ipv4))
                 {
-                    return true;
+                    return resolveValidationStack(validationStack);
                 }
             }
 
             var ipv6 = packet.Extract<IPv6Packet>();
             if (ipv6 != null)
             {
-                if (SourceIp != null && !SourceIp.Equals(ipv6.SourceAddress))
-                    return false;
+                if (SourceIp != null)
+                {
+                    validationStack.Push(() => SourceIp.Equals(ipv6.SourceAddress));
+                }
 
-                if (DestinationIp != null && DestinationIp.Equals(ipv6.DestinationAddress))
-                    return false;
+                if (DestinationIp != null)
+                {
+                    validationStack.Push(() => DestinationIp.Equals(ipv6.DestinationAddress));
+                }
 
                 if (RuleProtocol.Equals(Protocol.Ipv6))
                 {
-                    return true;
+                    return resolveValidationStack(validationStack);
                 }
             }
 
             var icmpv4 = packet.Extract<IcmpV4Packet>();
             if (icmpv4 != null && RuleProtocol.Equals(Protocol.Icmpv4))
             {
-                return (DestinationPort == 0 || (((ushort)icmpv4.TypeCode) / 256) == DestinationPort);
+                if (DestinationPort != null)
+                {
+                    validationStack.Push(() => (((ushort)icmpv4.TypeCode) / 256) == DestinationPort);
+                }
+
+                return resolveValidationStack(validationStack);
             }
 
             var icmpv6 = packet.Extract<IcmpV4Packet>();
             if (icmpv6 != null && RuleProtocol.Equals(Protocol.Icmpv6))
             {
-                return (DestinationPort == 0 || (((ushort)icmpv6.TypeCode) / 256) == DestinationPort);
+                if (DestinationPort != null)
+                {
+                    validationStack.Push(() => (((ushort)icmpv6.TypeCode) / 256) == DestinationPort);
+                }
+
+                return resolveValidationStack(validationStack);
             }
 
             var tcp = packet.Extract<TcpPacket>();
             if (tcp != null && RuleProtocol.Equals(Protocol.Tcp))
             {
-                if (SourcePort != AnyPort && !SourcePort.Equals(tcp.SourcePort))
-                    return false;
+                if (SourcePort != null)
+                {
+                    validationStack.Push(() => SourcePort.Equals(tcp.SourcePort));
+                }
 
-                return DestinationPort == AnyPort || DestinationPort.Equals(tcp.DestinationPort);
+                if (DestinationPort != null)
+                {
+                    validationStack.Push(() => DestinationPort.Equals(tcp.DestinationPort));
+                }
+
+                return resolveValidationStack(validationStack);
             }
 
             var udp = packet.Extract<UdpPacket>();
             if (udp != null && RuleProtocol.Equals(Protocol.Udp))
             {
-                if (SourcePort != AnyPort && !SourcePort.Equals(udp.SourcePort))
-                    return false;
+                if (SourcePort != null)
+                {
+                    validationStack.Push(() => SourcePort.Equals(udp.SourcePort));
+                }
 
-                return DestinationPort == AnyPort || DestinationPort.Equals(udp.DestinationPort);
+                if (DestinationPort != null)
+                {
+                    validationStack.Push(() => DestinationPort.Equals(udp.DestinationPort));
+                }
+
+                return resolveValidationStack(validationStack);
             }
 
             return false;